Add prototype compilation function
Compilation prototype is added. To use, click on "Compile Umple Model"
in the {Tools} menu in Orion when an umple model is open in the editor.

All output is done in the console as File I/O is limited in Orion when
using the plugin architecture. However, the compiler connects to
UmpleOnline without a hitch.

In a future commit, the compilation code will be added to a server-side
SDK plugin. This will mean that compilation will unfortunately be a
feature that our server must add on it's own.

Signed-off-by: Victoria Lacroix <victoria.a.lacroix@gmail.com>